+++ frontend/src/atoms/DatePicker.tsx
@@ -1,0 +1,48 @@
+import 'date-fns';
+import ja from "date-fns/locale/ja";
+import format from "date-fns/format";
+import { VFC, memo } from 'react';
+import DateFnsUtils from '@date-io/date-fns';
+import { MaterialUiPickersDate } from '@material-ui/pickers/typings/date';
+import {
+  MuiPickersUtilsProvider,
+  KeyboardDatePicker,
+} from '@material-ui/pickers';
+
+
+class ExtendedUtils extends DateFnsUtils {
+  getCalendarHeaderText(date:Date) {
+    return format(date, "yyyy MMM", { locale: this.locale });
+  }
+  getDatePickerHeaderText(date:Date) {
+    return format(date, "MMMd日", { locale: this.locale });
+  }
+}
+export type Props = {
+  onClickConfirm:()=>void
+  onChangeDate:(date: MaterialUiPickersDate)=>void
+  selectedDate: Date | string| null
+}
+
+export const DatePicker:VFC<Props>=memo((props)=> {
+  const { onClickConfirm, onChangeDate, selectedDate } = props
+
+  return (
+    <MuiPickersUtilsProvider locale={ja} utils={ExtendedUtils}>
+      <KeyboardDatePicker
+          margin="normal"
+          id="date"
+          label="日付選択"
+          okLabel="決定"
+          cancelLabel="キャンセル"
+          format="yyyy/MM/dd"
+          value={selectedDate}
+          onChange={onChangeDate}
+          KeyboardButtonProps={{
+            'aria-label': 'change date',
+          }}
+          onAccept={onClickConfirm}
+        />
+    </MuiPickersUtilsProvider>
+  );
+})
